requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

# LlamaIndex (single-version, no plugins)
llama-index==0.10.30

# Embeddings / LLM (Gemini – optional)
google-generativeai==0.3.2

# Utils
nltk==3.8.1





